run postgres cases in new db
